+++ bin/node/src/commands/store.rs
@@ -1,31 +1,17 @@
-use std::{
-    fs::File,
-    io::Write,
-    path::{Path, PathBuf},
-    time::{SystemTime, UNIX_EPOCH},
-};
+use std::path::{Path, PathBuf};
 
 use anyhow::Context;
-use miden_lib::{AuthScheme, account::faucets::create_basic_fungible_faucet, utils::Serializable};
-use miden_node_store::{GenesisState, Store};
-use miden_node_utils::{crypto::get_rpo_random_coin, grpc::UrlExt};
-use miden_objects::{
-    Felt, ONE,
-    account::{Account, AccountFile, AuthSecretKey},
-    asset::TokenSymbol,
-    crypto::dsa::rpo_falcon512::SecretKey,
+use miden_node_store::{
+    Store,
+    genesis::config::{AccountFileWithName, GenesisConfig},
 };
-use rand::{Rng, SeedableRng};
-use rand_chacha::ChaCha20Rng;
+use miden_node_utils::grpc::UrlExt;
 use url::Url;
 
 use super::{
     ENV_DATA_DIRECTORY, ENV_STORE_BLOCK_PRODUCER_URL, ENV_STORE_NTX_BUILDER_URL, ENV_STORE_RPC_URL,
 };
-use crate::commands::ENV_ENABLE_OTEL;
-
-/// The default filepath for the genesis account.
-const DEFAULT_ACCOUNT_PATH: &str = "account.mac";
+use crate::commands::{ENV_ENABLE_OTEL, ENV_GENESIS_CONFIG_FILE};
 
 #[allow(clippy::large_enum_variant, reason = "single use enum")]
 #[derive(clap::Subcommand)]
@@ -40,9 +26,12 @@ pub enum StoreCommand {
         /// Directory in which to store the database and raw block data.
         #[arg(long, env = ENV_DATA_DIRECTORY, value_name = "DIR")]
         data_directory: PathBuf,
-        // Directory to write the account data to.
+        /// Directory to write the account data to.
         #[arg(long, value_name = "DIR")]
         accounts_directory: PathBuf,
+        /// Use the given configuration file to construct the genesis state from.
+        #[arg(long, env = ENV_GENESIS_CONFIG_FILE, value_name = "GENESIS_CONFIG")]
+        genesis_config_file: Option<PathBuf>,
     },
 
     /// Starts the store component.
@@ -79,8 +68,12 @@ impl StoreCommand {
     /// Executes the subcommand as described by each variants documentation.
     pub async fn handle(self) -> anyhow::Result<()> {
         match self {
-            StoreCommand::Bootstrap { data_directory, accounts_directory } => {
-                Self::bootstrap(&data_directory, &accounts_directory)
+            StoreCommand::Bootstrap {
+                data_directory,
+                accounts_directory,
+                genesis_config_file,
+            } => {
+                Self::bootstrap(&data_directory, &accounts_directory, genesis_config_file.as_ref())
             },
             StoreCommand::Start {
                 rpc_url,
@@ -138,79 +131,35 @@ impl StoreCommand {
         .context("failed while serving store component")
     }
 
-    fn bootstrap(data_directory: &Path, accounts_directory: &Path) -> anyhow::Result<()> {
-        // Generate the genesis accounts.
-        let account_file =
-            Self::generate_genesis_account().context("failed to create genesis account")?;
+    fn bootstrap(
+        data_directory: &Path,
+        accounts_directory: &Path,
+        maybe_genesis_config: Option<&PathBuf>,
+    ) -> anyhow::Result<()> {
+        let config = maybe_genesis_config
+            .map(|genesis_config| {
+                let toml_str = fs_err::read_to_string(genesis_config)?;
+                let config = GenesisConfig::read_toml(toml_str.as_str())
+                    .context(format!("Read from file: {}", genesis_config.display()))?;
+                Ok::<_, anyhow::Error>(config)
+            })
+            .transpose()?
+            .unwrap_or_default();
 
-        // Write account data to disk (including secrets).
-        //
-        // Without this the accounts would be inaccessible by the user.
-        // This is not used directly by the node, but rather by the owner / operator of the node.
-        let filepath = accounts_directory.join(DEFAULT_ACCOUNT_PATH);
-        File::create_new(&filepath)
-            .and_then(|mut file| file.write_all(&account_file.to_bytes()))
-            .with_context(|| {
-                format!("failed to write data for genesis account to file {}", filepath.display())
-            })?;
+        let (genesis_state, secrets) = config.into_state()?;
+        // Write the accounts to disk
+        for item in secrets.as_account_files(&genesis_state) {
+            let AccountFileWithName { account_file, name } = item?;
+            let accountpath = accounts_directory.join(name);
+            // do not override existing keys
+            fs_err::OpenOptions::new()
+                .create_new(true)
+                .write(true)
+                .open(&accountpath)
+                .context("key file already exists")?;
+            account_file.write(accountpath)?;
+        }
 
-        // Bootstrap the store database.
-        let version = 1;
-        let timestamp = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .expect("current timestamp should be greater than unix epoch")
-            .as_secs()
-            .try_into()
-            .expect("timestamp should fit into u32");
-        let genesis_state = GenesisState::new(vec![account_file.account], version, timestamp);
         Store::bootstrap(genesis_state, data_directory)
-    }
-
-    fn generate_genesis_account() -> anyhow::Result<AccountFile> {
-        let mut rng = ChaCha20Rng::from_seed(rand::random());
-        let secret = SecretKey::with_rng(&mut get_rpo_random_coin(&mut rng));
-
-        // Calculate the max supply of the token.
-        let decimals = 6u8;
-        let base_unit = 10u64.pow(u32::from(decimals));
-        let max_supply = 100_000_000_000u64 * base_unit;
-        let max_supply = Felt::try_from(max_supply).expect("max supply is less than field modulus");
-
-        // Create the faucet.
-        let (account, account_seed) = create_basic_fungible_faucet(
-            rng.random(),
-            TokenSymbol::try_from("MIDEN").expect("MIDEN is a valid token symbol"),
-            decimals,
-            max_supply,
-            miden_objects::account::AccountStorageMode::Public,
-            AuthScheme::RpoFalcon512 { pub_key: secret.public_key() },
-        )?;
-
-        // Force the account nonce to 1.
-        //
-        // By convention, a nonce of zero indicates a freshly generated local account that has yet
-        // to be deployed. An account is deployed onchain along with its first transaction which
-        // results in a non-zero nonce onchain.
-        //
-        // The genesis block is special in that accounts are "deplyed" without transactions and
-        // therefore we need bump the nonce manually to uphold this invariant.
-        let (id, vault, sorage, code, _) = account.into_parts();
-        let updated_account = Account::from_parts(id, vault, sorage, code, ONE);
-
-        Ok(AccountFile::new(
-            updated_account,
-            Some(account_seed),
-            vec![AuthSecretKey::RpoFalcon512(secret)],
-        ))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::StoreCommand;
-
-    #[test]
-    fn generate_genesis_account_no_panic() {
-        let _account = StoreCommand::generate_genesis_account().unwrap();
     }
 }